Add `unassign` method to a conversation allowing anyone to be assigned
A support issue was pointed out that via the PHP library, you could not unassign a user
from a conversation. The API docs specify to send a `null` value for `owner` to unassign.
This was not possible in this library. To add this functionality and be as minimally invasive
as possible, a new property and method was added (`unassign`). This is a simple executional method.
When invoked, the owner is set to null and a property of `unassigned` is set to true. In the
`getObjectVars` method, a check was added that if the owner is null and `unassigned` is `true`,
to set the `vars['owner']` to null thus allowing an unassignment.

A new test (`UnassignConversationTest`) was added to verify the expected results.